pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["t_ishikawa <tissi0708@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
apache-airflow = "^1.10.9"
six = "^1.14.0"
attrs = "^19.3.0"
urllib3 = "^1.25.8"
chardet = "^3.0.4"
certifi = "^2019.11.28"
lockfile = "^0.12.2"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
